operator/controller: Split reconcile loop per resource
